Add feature: change prefix or copy mode color bg
The user can now highlight the text displayed when the prefix key is
pressed and / or when it is in copy mode by changing the background
colors associated to "PREFIX_PRESSED" and "COPY".

To do so they can modify the value of two new variables as shown in
the README file.

To ensure that those two backgrounds wrap their associated text and
nothing else (adjacent white spaces and the separator text) two white
spaces have been moved from being printed with 'echo -n " "' to being
included in the 'separator' and 'right_edge_char' variables.